blood_bridge_backend/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-authtoken==1.1.0
django-cors-headers==4.0.0
